refactor: moved modal outside buttons
